Consolidate simulation and solution state-choice-space functions
